refactor(structure): Consistency between 'pages' files & routes
