+++ src/pages/Tasks/index.tsx
@@ -1,16 +1,16 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
-import AssignedTasks from './assigned';
+import TaskDetails from './details';
+import EditTask from './edit';
+import AssignedTasks from './main';
 import NewTask from './new';
-import TaskDetail from './task-detail';
-import UpdateTask from './update';
 
 const Tasks = () => {
   return (
     <Routes>
       <Route path='/' element={<AssignedTasks />} />
       <Route path='/:projectId/create' element={<NewTask />} />
-      <Route path='/:id' element={<TaskDetail />} />
-      <Route path='/update/:id' element={<UpdateTask />} />
+      <Route path='/:id' element={<TaskDetails />} />
+      <Route path='/edit/:id' element={<EditTask />} />
       <Route path='*' element={<Navigate to='/404' />} />
     </Routes>
   );
